requirements.txt
mss==3.3.2
numpy==1.15.4
opencv-python==3.4.4.19
pdfkit==0.6.1
Pillow==5.3.0
pkg-resources==0.0.0
